Separate user input from display text
When performing an operation (e.g. 2 + 3), the display will be cleared
after the user inputs the second operand. Thus, it is necessary to
separate what is shown on the display from the user's input.

Consequently, separated the logic for updating the user input and
display text, and renamed the variable for storing the user's input.